% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalan_parties_colours.R
\docType{data}
\name{catparties_palette}
\alias{catparties_palette}
\title{catparties_palette}
\format{
An object of class \code{character} of length 19.
}
\usage{
catparties_palette
}
\description{
Official colours for Catalan political parties 2010-2022
Colors oficials dels partits polítics catalans 2010-2022
JxCat es refereix a la Coalició JxCat que es va presentar a les eleccions del 2017 (rosat). Junts, al partit polític posterior (turquesa)
JxCat refers to the JxCat coalition that ran on the 2017 election (pink). Junts, to its successor party (turquoise)
#' Names are usually given in abbreviated form unless not frequently used in daily speech. It should be trivial to change it.
}
\keyword{datasets}
