% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverging_colours.R
\docType{data}
\name{contrasting_palette}
\alias{contrasting_palette}
\title{diverging_palette}
\format{
An object of class \code{character} of length 6.
}
\usage{
contrasting_palette
}
\description{
this is a palette with diverging colours. Relatively muted brown, green, and blue tones.
}
\keyword{datasets}
