% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubois_colours.R
\docType{data}
\name{dubois_palette_long}
\alias{dubois_palette_long}
\title{dubois_palette_long}
\format{
An object of class \code{character} of length 8.
}
\usage{
dubois_palette_long
}
\description{
This palette collects the main colours used by WEB DuBois.
It can be used as a contrasting palette but they represent the order used by WEB DuBois in quantitative choropleth maps and can be used as such.
Use \code{dubois_palette_short()} when less colours are needed to guarantee that there is enough contrast.
}
\keyword{datasets}
