% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casabloc_colours.R
\docType{data}
\name{palette_bloc}
\alias{palette_bloc}
\title{bloc_colours}
\format{
An object of class \code{character} of length 4.
}
\usage{
palette_bloc
}
\description{
a browns palette inspired by the colours of Casa Bloc. Can function both as continuous and as categorical.
}
\keyword{datasets}
