% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steel_colours.R
\docType{data}
\name{steel_palette}
\alias{steel_palette}
\title{steel_palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
steel_palette
}
\description{
a steel blue palette.
}
\keyword{datasets}
