% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dubois_minimal.R
\name{theme_dubois_classic}
\alias{theme_dubois_classic}
\title{theme_dubois_classic}
\usage{
theme_dubois_classic(
  base_size = 11,
  base_family = "Roboto Condensed",
  base_line_size = base_size/170,
  base_rect_size = base_size/170,
  legend_position = NULL,
  bg_fill = "antiquewhite1"
)
}
\arguments{
\item{base_size}{font size to use}

\item{base_family}{font family to use}

\item{base_line_size}{equal to the param in theme_minimal}

\item{legend_position}{equal to the param in theme}

\item{bg_fill}{background colour}

\item{base_grid_size}{equal to the param in ggplot2::theme_minimal}
}
\description{
ggplot theme in the style of WEB DuBois without any grid
}
